# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/roleActions": {
            "get": {
                "tags": [
                    "RoleAction"
                ],
                "summary": "get all role's actions",
                "operationId": "e39d95c1ebc122b9e253ce0eb1f6e64d",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success,return array of action objects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        },
        "/roleAttachAction": {
            "post": {
                "tags": [
                    "RoleAction"
                ],
                "summary": "attach a action to role",
                "operationId": "bd49ccf79d99d0dee1df4ee9f2342102",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "action_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        },
        "/roleDetachAction": {
            "post": {
                "tags": [
                    "RoleAction"
                ],
                "summary": "detach a action of role",
                "operationId": "ae0a7692d1cb9bfed2f334583ef2b527",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "action_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        },
        "/roles": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "get roles",
                "operationId": "getRoles",
                "responses": {
                    "200": {
                        "description": "success,return array of roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            },
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "store a role",
                "operationId": "b47f38c8385567ad7a473d82fb6ebf63",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "role name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Whether a role is active or not",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success,return object of role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        },
        "/roles/{role_id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "get a role by id",
                "operationId": "4dbe05b743da86873c230cb8f8bd2bc7",
                "responses": {
                    "200": {
                        "description": "success,return object of role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            },
            "put": {
                "tags": [
                    "Role"
                ],
                "summary": "update a role",
                "operationId": "f5dec4618fb663a4e726bbb87db517be",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "role name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Whether a role is active or not",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Role"
                ],
                "summary": "delete a role",
                "operationId": "bffe2b04e26262a7c957f15a691ccdcd",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    }
}